metadata: add duration

+++ common/src/main/java/com/liskovsoft/smartyoutubetv2/common/app/models/data/SampleMediaItem.java
@@ -184,7 +184,7 @@ public final class SampleMediaItem implements MediaItem {
     }
 
     @Override
-    public int getDurationMs() {
+    public long getDurationMs() {
         return 0;
     }
